[CTTSK-3] refactor: mapper 삭제 및 패키지 이동
