+++ src/main/java/com/cliptripbe/feature/bookmark/application/BookmarkService.java
@@ -11,8 +11,8 @@ import com.cliptripbe.feature.bookmark.dto.request.UpdateBookmarkRequestDto;
 import com.cliptripbe.feature.bookmark.dto.response.BookmarkInfoResponseDto;
 import com.cliptripbe.feature.bookmark.dto.response.BookmarkListResponseDto;
 import com.cliptripbe.feature.bookmark.infrastructure.BookmarkRepository;
-import com.cliptripbe.feature.place.api.dto.PlaceInfoRequestDto;
-import com.cliptripbe.feature.place.api.dto.response.PlaceListResponseDto;
+import com.cliptripbe.feature.place.dto.request.PlaceInfoRequestDto;
+import com.cliptripbe.feature.place.dto.response.PlaceListResponseDto;
 import com.cliptripbe.feature.place.application.PlaceService;
 import com.cliptripbe.feature.place.domain.entity.Place;
 import com.cliptripbe.feature.place.domain.entity.PlaceTranslation;
