+++ src/main/java/com/cliptripbe/feature/bookmark/dto/request/UpdateBookmarkRequestDto.java
@@ -1,6 +1,6 @@
 package com.cliptripbe.feature.bookmark.dto.request;
 
-import com.cliptripbe.feature.place.api.dto.PlaceInfoRequestDto;
+import com.cliptripbe.feature.place.dto.request.PlaceInfoRequestDto;
 import java.util.List;
 
 public record UpdateBookmarkRequestDto(
