+++ src/main/java/com/cliptripbe/feature/place/application/PlaceTranslationService.java
@@ -2,7 +2,6 @@ package com.cliptripbe.feature.place.application;
 
 import static com.cliptripbe.feature.user.domain.type.Language.ENGLISH;
 
-import com.cliptripbe.feature.place.api.dto.PlaceDto;
 import com.cliptripbe.feature.place.domain.entity.Place;
 import com.cliptripbe.feature.place.domain.entity.PlaceTranslation;
 import com.cliptripbe.feature.place.domain.vo.TranslationInfo;
