+++ src/main/java/com/cliptripbe/feature/place/dto/request/PlaceSearchByKeywordRequestDto.java
@@ -1,4 +1,4 @@
-package com.cliptripbe.feature.place.api.dto.request;
+package com.cliptripbe.feature.place.dto.request;
 
 import jakarta.validation.constraints.Max;
