+++ src/main/java/com/cliptripbe/feature/schedule/dto/request/CreateScheduleRequestDto.java
@@ -1,6 +1,6 @@
 package com.cliptripbe.feature.schedule.dto.request;
 
-import com.cliptripbe.feature.place.api.dto.PlaceInfoRequestDto;
+import com.cliptripbe.feature.place.dto.request.PlaceInfoRequestDto;
 import java.util.List;
 
 public record CreateScheduleRequestDto(
